Move Designator to expr package
