+++ FortranAst/src/pt/up/fe/specs/fortran/ast/nodes/stmt/AssignmentStmt.java
@@ -3,7 +3,7 @@ package pt.up.fe.specs.fortran.ast.nodes.stmt;
 import org.suikasoft.jOptions.Interfaces.DataStore;
 import pt.up.fe.specs.fortran.ast.nodes.FortranNode;
 import pt.up.fe.specs.fortran.ast.nodes.expr.Expr;
-import pt.up.fe.specs.fortran.ast.nodes.variable.Designator;
+import pt.up.fe.specs.fortran.ast.nodes.expr.Designator;
 
 import java.util.Collection;
