+++ FortranAst/src/pt/up/fe/specs/fortran/ast/nodes/variable/DataRef.java
@@ -4,6 +4,7 @@ import org.suikasoft.jOptions.Datakey.DataKey;
 import org.suikasoft.jOptions.Datakey.KeyFactory;
 import org.suikasoft.jOptions.Interfaces.DataStore;
 import pt.up.fe.specs.fortran.ast.nodes.FortranNode;
+import pt.up.fe.specs.fortran.ast.nodes.expr.Designator;
 
 import java.util.Collection;
